# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Get admin dashboard totals
      responses:
        "200":
          description: Admin overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List customer token accounts and usage
      responses:
        "200":
          description: Customer accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/tokens:
    post:
      operationId: createAdminToken
      tags: [admin]
      summary: Issue a customer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenInput"
      responses:
        "201":
          description: Token created; plaintext token is returned once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedTokenResponse"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/tokens/{tokenId}/reveal:
    post:
      operationId: revealAdminToken
      tags: [admin]
      summary: Reveal a token through an audited encrypted recovery action
      parameters:
        - $ref: "#/components/parameters/TokenId"
      responses:
        "200":
          description: Revealed token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedTokenResponse"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Token not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/tokens/{tokenId}/revoke:
    post:
      operationId: revokeAdminToken
      tags: [admin]
      summary: Revoke a customer token
      parameters:
        - $ref: "#/components/parameters/TokenId"
      responses:
        "204":
          description: Token revoked
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{userId}/status:
    patch:
      operationId: updateAdminUserStatus
      tags: [admin]
      summary: Restrict or restore a customer account
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserStatusInput"
      responses:
        "200":
          description: Updated customer account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List orders for manual payment review
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderListResponse"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{orderId}/status:
    patch:
      operationId: updateAdminOrderStatus
      tags: [admin]
      summary: Mark an order pending, successful, or failed
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/settings/payment:
    get:
      operationId: getAdminPaymentSettings
      tags: [admin]
      summary: Get BTC payment configuration
      responses:
        "200":
          description: Payment settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSettings"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdminPaymentSettings
      tags: [admin]
      summary: Update BTC payment configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSettingsInput"
      responses:
        "200":
          description: Updated payment settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSettings"
        "401":
          description: Admin session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
...
    TokenId:
      name: tokenId
      in: path
      required: true
      schema:
        type: string
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, restricted, deleted]
        slipLimit:
          type: number
        slipsUsed:
          type: number
        designLimit:
          type: number
        designsUsed:
          type: number
        pdfPrintCount:
          type: number
        browserPrintCount:
          type: number
        lockoutUntil:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, status, slipLimit, slipsUsed, designLimit, designsUsed, pdfPrintCount, browserPrintCount, lockoutUntil, createdAt]
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/UserSummary"
        - type: object
          properties:
            tokenId:
              type: string
            tokenStatus:
              type: string
              enum: [active, revoked]
          required: [tokenId, tokenStatus]
...
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        planId:
          type: string
          enum: [starter, pro, enterprise]
        name:
          type: string
        email:
          type: string
        paymentMethod:
          type: string
          enum: [card_manual, bitcoin]
        status:
          type: string
          enum: [pending, successful, failed]
        amountCents:
          type: number
        currency:
          type: string
        acceptedPaymentTerms:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, orderNumber, planId, name, email, paymentMethod, status, amountCents, currency, acceptedPaymentTerms, createdAt, updatedAt]
...
    AdminOverview:
      type: object
      properties:
        totalUsers:
          type: number
        activeUsers:
          type: number
        restrictedUsers:
          type: number
        totalPdfPrints:
          type: number
        totalBrowserPrints:
          type: number
        pendingOrders:
          type: number
        successfulOrders:
          type: number
        failedOrders:
          type: number
      required: [totalUsers, activeUsers, restrictedUsers, totalPdfPrints, totalBrowserPrints, pendingOrders, successfulOrders, failedOrders]
    AdminUserListResponse:
      type: array
      items:
        $ref: "#/components/schemas/AdminUser"
    CreateTokenInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
        planId:
          type: string
          enum: [starter, pro, enterprise]
      required: [name, email, planId]
    IssuedTokenResponse:
      type: object
      properties:
        tokenId:
          type: string
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserSummary"
      required: [tokenId, token, user]
    UserStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [active, restricted, deleted]
      required: [status]
    AdminOrderListResponse:
      type: array
      items:
        $ref: "#/components/schemas/Order"
    OrderStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, successful, failed]
      required: [status]
    PaymentSettings:
      type: object
      properties:
        bitcoinWallet:
          type: string
        bitcoinQrUrl:
          type: ["string", "null"]
        bitcoinInstructions:
          type: string
      required: [bitcoinWallet, bitcoinQrUrl, bitcoinInstructions]
    PaymentSettingsInput:
      allOf:
        - $ref: "#/components/schemas/PaymentSettings"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]