# lib/api-zod/src/generated/types/createTokenInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTokenInputPlanId } from './createTokenInputPlanId';

export interface CreateTokenInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  email: string;
  planId: CreateTokenInputPlanId;
}