# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderPlanId } from './orderPlanId';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  orderNumber: string;
  planId: OrderPlanId;
  name: string;
  email: string;
  paymentMethod: OrderPaymentMethod;
  status: OrderStatus;
  amountCents: number;
  currency: string;
  acceptedPaymentTerms: boolean;
  createdAt: string;
  updatedAt: string;
}