# lib/api-zod/src/generated/types/paymentSettingsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentSettings } from './paymentSettings';

export type PaymentSettingsInput = PaymentSettings;